+++ EQLogParser/src/control/util/TriggerUtil.cs
@@ -78,8 +78,10 @@ namespace EQLogParser
         toTrigger.PreviousPattern = TextUtils.Trim(fromTrigger.PreviousPattern);
         toTrigger.EndEarlyPattern = TextUtils.Trim(fromTrigger.EndEarlyPattern);
         toTrigger.EndEarlyPattern2 = TextUtils.Trim(fromTrigger.EndEarlyPattern2);
+        toTrigger.EndEarlyPattern3 = TextUtils.Trim(fromTrigger.EndEarlyPattern3);
         toTrigger.EndUseRegex = fromTrigger.EndUseRegex;
         toTrigger.EndUseRegex2 = fromTrigger.EndUseRegex2;
+        toTrigger.EndUseRegex3 = fromTrigger.EndUseRegex3;
         toTrigger.EndEarlyRepeatedCount = fromTrigger.EndEarlyRepeatedCount;
         toTrigger.WorstEvalTime = fromTrigger.WorstEvalTime;
         toTrigger.ResetDurationSeconds = fromTrigger.ResetDurationSeconds;
@@ -115,6 +117,8 @@ namespace EQLogParser
         if (toTrigger is TriggerPropertyModel toModel)
         {
           toModel.TriggerActiveBrush = UiUtil.GetBrush(fromTrigger.ActiveColor, false);
+          toModel.TriggerIdleBrush = UiUtil.GetBrush(fromTrigger.IdleColor, false);
+          toModel.TriggerResetBrush = UiUtil.GetBrush(fromTrigger.ResetColor, false);
           toModel.TriggerFontBrush = UiUtil.GetBrush(fromTrigger.FontColor, false);
 
           var (textItems, timerItems) = await GetOverlayItems(toModel.SelectedOverlays);
@@ -141,6 +145,8 @@ namespace EQLogParser
         else if (fromTrigger is TriggerPropertyModel fromModel)
         {
           toTrigger.ActiveColor = fromModel.TriggerActiveBrush?.Color.ToHexString();
+          toTrigger.IdleColor = fromModel.TriggerIdleBrush?.Color.ToHexString();
+          toTrigger.ResetColor = fromModel.TriggerResetBrush?.Color.ToHexString();
           toTrigger.FontColor = fromModel.TriggerFontBrush?.Color.ToHexString();
 
           var selectedOverlays = fromModel.SelectedTextOverlays.Where(item => item.IsChecked).Select(item => item.Value).ToList();